src/forge_requirements_builder/requirements.txt
# Forge Requirements Builder - Dependencies

# Core LangGraph & AI
langgraph>=0.2.0
langchain-openai>=0.1.0
openai>=1.0.0

# State Management & Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# UI/API
streamlit>=1.30.0
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6

# Document Processing
pypdf>=3.0.0
python-docx>=1.0.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0

# Development
black>=23.0.0
ruff>=0.1.0
mypy>=1.0.0
